Do not take remote-only parents when splitting passage
When splitting a passage to sentences or paragraphs, we find the set of nodes participating in the sentence/paragraph by going up from the words until we reach the root.
Remove edges will no longer be taken into account in this traversal, to avoid cases like passage 494, unit 1.19, where there are two remotes in the first sentence and one real child in the second sentence. If we include 1.19 in the first sentence's units, it will have only remote children, which is not allowed.