Re-enable linkage edges in conversion from/to SDP
The problem was that often cycles are created by linkage edges, because the predicate has a Linker edge going to the link relation, which in turn has a Link Argument edge going to the predicate.
Cycles are now detected and removed by deleting remote and Linker edges, or, if there are none, any other edge, until no cycle exists.
Performance is now F1~.9 on regular edges and F1~.5 on remote edges when converting to SDP and back.