Option to disable the Swap operation
If disabled, it means that the theoretical coverage is no longer 100%, that is, the oracle is no longer able to parse all the passages in the corpus. However, the model is simpler which might yield better performance anyway, since the percentage of discontinuous units in the English corpus (which is all there is right now) is low.